Decouple library from Arduino SPI functions
Instead, pass in SPI function pointers to be used inside the SD driver.
We now assume that the user of the library has properly initialized the
SPI module. The user is also responsible for choosing the speed at which
the SPI module runs. This requires a bit more work on the part of the
library user, but it means that this library can now be used on any
microcontroller that supports SPI.

One of the examples has been updated. The previously hidden SPI
functions are now defined in that sketch. The SD card class takes in
pointers to those functions. The rest of the examples will be deleted
in a subsequent commit since I don't want to update them and the one I
updated has enough information.