Break out private functions from SdCard class
Remove the private SD handler functions and turn them into static helper
functions. These now operate on a pointer to an SD card object. Since we
can't pass "this" to a non-member function as an argument, we
temporarily get around this, we maintain a static pointer to an SD card
object. This is set to "this" upon construction. This is what gets
passed to the helper functions. This effectively means that we can't
safely have multiple SD card objects floating around, but this is very
very temporary so it's fine.